ajouter la fonctionnalité d'ajout

+++ supprimerService.php
@@ -1,16 +1,21 @@
 <?php
-require_once("connexiondb.php");
+session_start();
+if(isset($_SESSION['user'])){
+        require_once("connexiondb.php");
 
-$codes = isset($_GET['codes']) ? $_GET['codes'] : 0;
-
-
-    $requete = "DELETE FROM Service WHERE CodeS = ?";
-    $params = array($codes);
-
-    $resultat = $pdo->prepare($requete);
-    $resultat->execute($params);
+        $codes = isset($_GET['codes']) ? $_GET['codes'] : 0;
 
 
-header('Location: service.php');
+            $requete = "DELETE FROM Service WHERE CodeS = ?";
+            $params = array($codes);
+
+            $resultat = $pdo->prepare($requete);
+            $resultat->execute($params);
+
+
+        header('Location: service.php');
+    }else {
+        header('location:authentification.php');
+     }
 
 ?>
